initial gas optimization suggestions
Notes:

BidTicket.sol:
Recommendation - Nest the if statement conditions so that calls from the harvest and market contracts can bypass the owner check.
Recommendation - Remove the data parameter from the mint and mintBatch functions, pass an empty bytes array to _mint and _mintBatch

Harvest.sol:
supportsInterface is not always reliable and gas cost to make the call can vary from contract to contract
Recommendation - use count == 0 for ERC721 tokens, count > 0 for ERC1155 tokens OR split batchTransfer into two functions, one for ERC721 and one for ERC1155

When calling batchTransfer with multiple tokens that do not have a custom price, the defaultPrice storage variable gets loaded multiple times.
Recommendation - load defaultPrice to a memory variable and pass to _getPrice to avoid extra storage reads

MaxTokensPerTxReached check is applied to every iteration of the for loop in batchTransfer.
Recommendation - move the check to after the loop to check once, wallet UI will advise user that the transaction is expected to revert

Checking for token approvals onchain.
Recommendation - remove token approval check, wallet UI will advise user that the transaction is expected to revert due to lack of approvals

Market.sol:
supportsInterface is not always reliable and gas cost to make the call can vary from contract to contract
Recommendation - split startAuction into two functions, one for ERC721 and one for ERC1155, and add auctionType to Auction struct to determine type for claim

Auction struct is defined in a way that does not efficiently pack storage variables
Recommendation - Move highestBidder and highestBid after withdrawn so that tokenAddress, endTime, tokenCount, claimed and withdrawn are packed into a single slot. Reduces storage for auctions from 4 slots (+ tokens) to 3 slots (+ tokens)

Token amounts are unnecessary for ERC721 auctions.
Recommendation - do not store amounts for ERC721 tokens, auctionType can be used to return amount = 1 in the getAuctionTokens function for ERC721 token auctions

Recommendation - Cache the storage mappings for auction.tokenIds, auction.amounts, and auctionTokens in startAuction, validate and transfer token functions

Recommendation - Cache auction.tokenCount in all for loops that iterate over auction tokens

Recommendation - Cache tokenAddress in transfer functions

Solidity uses 32 byte slots on the stack, in memory and in storage, type casting variables down below uint256 incurs validation overhead
Recommendation - Change uint8 mapping keys to uint256

For auctions to be able to be claimed the barn needs to set approval for the market contract to transfer tokens. There is currently no method for the barn to be able to automatically set approvals. Assuming the barn is going to be a smart contract, it should have a method that's callable by the market contract to set approval to the market contract.
Recommendation - create the barn contract with a function that can be called to set approval to Market if the token approval is not currently set, the barn will need to implement ERC1155Receiver to receive 1155 tokens